refactor: 패키지 구조 수정 및 테스트 주석 추가
